201106 ft_printf

+++ 01_ft_printf/srcs/ft_print_float.c
@@ -6,7 +6,7 @@
 /*   By: hyunlee <hyunlee@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/05 20:16:10 by hyunlee           #+#    #+#             */
-/*   Updated: 2020/11/05 21:16:03 by hyunlee          ###   ########.fr       */
+/*   Updated: 2020/11/06 19:37:43 by hyunlee          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_print_float(t_set *set)
 {
-	long double num;
+	double num;
